Attempt to read in new datasets
Wow - it took a long time to figure out just how to read in the data at all, and I'm still not finished, but this seems like a good time to take a break. 

As a next step, can anyone figure out how to translate m_credits.cast[0] to a list of names?
My understanding (just based only seeing "buzz lightyear" credited) is that the first
movie on the credits list is Toy Story. The cast field for Toy Story returns
a list of cast members, along with their genders and some other info. Most of this
information is irrelevant to us. So far, I tried a few things (including the JSON normalize code from our last hw) but haven't been successful.