Cleaned up comments
Also started quality checks and data cleaning process, but there's still more to do.

Note that this version of the code uses a right-join with the Bechdel test. As shown in line 343, currently using a right-join so that we can see which movies DO have Bechdel results. Eventually we'll want to also include movies that don't have Bechdel results. The reason that I started with a right-join was to make the cleaning task more manageable to start.