Left join & limit to 1990-2020
This is my follow-up work after our group meeting. I will revisit this tomorrow to see if the quality checks are complete, since changing the year variable to integer broke the mergekey, thus requiring a lot of troubleshooting. Hopefully everything is fixed, but visiting it with fresh eyes tomorrow will be important.