Updated to change CSV name back to what Betty used
Since the API work took SO many hours, we are going back a few steps in the master code (ex. changing NA values to 'Missing') so that this file creates the same version of movietests_20200227_v2.csv that is used in the next step.

KR note: Still need to use Excel comparison feature to make 100% sure that this creates the same version of the file that Betty used for her work.