One more try/catch + separated cast and crew
I broke the app when I entered an invalid number, so I added a try/catch block for this. Also separated cast and crew into separate menu options since both contain two figures now. One final change I added after having Erika try using the app is that I fixed the formatting on the part of the app that shows sample size. This should be final now, but if anyone else can check on their computer that'd be lovely.